Define app helper method
