+++ test/weather_web/views/weather_view_test.exs
@@ -1,3 +1,3 @@
-defmodule WeatherWeb.PageViewTest do
+defmodule WeatherWeb.WeatherViewTest do
   use WeatherWeb.ConnCase, async: true
 end
